+++ lib/result/error.dart
@@ -1,6 +1,12 @@
+import 'package:meta/meta.dart';
+import 'dart:core';
+
 /// BaseError defines the Error interface, which can be used
 /// to implement custom error types.
+@immutable
 abstract class BaseError {
+  const BaseError();
+
   String error();
 
   @override
@@ -9,6 +15,7 @@ abstract class BaseError {
 
 /// Error is the standard error class using a string as
 /// the underlaying type.
+@immutable
 class Error implements BaseError {
   final String _err;
 
@@ -19,6 +26,7 @@ class Error implements BaseError {
 }
 
 /// isErrorType returns true if v is of error type T.
+@Deprecated('use "err is Errortype" instead.')
 bool isErrorType<T extends BaseError>(dynamic v) {
   // TODO: Check why "if(v is T) return true;" doesn't always work.
   try {
@@ -33,4 +41,5 @@ bool isErrorType<T extends BaseError>(dynamic v) {
 ///
 /// Any object implementing the BaseError interface
 /// is considered to be an error.
+@Deprecated('use "err is BaseError" instead.')
 bool isError(dynamic v) => v is BaseError;