+++ lib/usecase/usecase.dart
@@ -1,6 +1,10 @@
-import 'package:appydart/appydart.dart';
-export 'package:appydart/appydart.dart';
+import 'package:appydart/result/result.dart';
 
-abstract class Usecase<T, A> {
+abstract class BaseUsecase<T, A> {
+  Future<T> run(A args);
+}
+
+abstract class Usecase<T, A> extends BaseUsecase<Result<T>, A> {
+  @override
   Future<Result<T>> run(A args);
 }